requirements.txt
Flask==3.0.0
snowflake-connector-python==3.6.0
pandas==2.1.4
Flask-WTF==1.2.1
WTForms==3.1.1
plotly==5.18.0
python-dotenv==1.0.0
configparser==6.0.0